rag-api/requirements.txt
# Core Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# OpenAI
openai>=1.0.0

# Vector Database
chromadb>=0.4.0

# Google Cloud (for Firestore and BigQuery)
google-cloud-firestore>=2.11.0
google-cloud-bigquery>=3.11.0

# HTTP Client
aiohttp>=3.9.0
httpx>=0.25.0

# Audio/Video Processing
mutagen>=1.47.0
Pillow>=10.0.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0

# Monitoring (optional)
psutil>=5.9.0

# Other utilities
python-multipart>=0.0.6
PyJWT>=2.8.0
bleach>=6.0.0
redis>=5.0.0

# Testing
httpx>=0.25.0
colorama>=0.4.6
pytest>=7.4.0
pytest-asyncio>=0.21.0
